* [core]: fix compiler warnings
gc -Wextra signed-unsigned comare, unused, ...

+++ vsnprintf.c
@@ -132,7 +132,7 @@ int fl_vsnprintf(char* buffer, size_t bufsize, const char* format, va_list ap) {
 	case 'e' :
 	case 'f' :
 	case 'g' :
-	  if ((width + 2) > sizeof(temp)) break;
+	  if ((size_t)(width + 2) > sizeof(temp)) break;
 
 	  sprintf(temp, tformat, va_arg(ap, double));
 
@@ -157,7 +157,7 @@ int fl_vsnprintf(char* buffer, size_t bufsize, const char* format, va_list ap) {
 	case 'o' :
 	case 'u' :
 	case 'x' :
-	  if ((width + 2) > sizeof(temp)) break;
+	  if ((size_t)(width + 2) > sizeof(temp)) break;
 
 	  sprintf(temp, tformat, va_arg(ap, int));
 
@@ -175,7 +175,7 @@ int fl_vsnprintf(char* buffer, size_t bufsize, const char* format, va_list ap) {
 	  break;
 	    
 	case 'p' : /* Pointer value */
-	  if ((width + 2) > sizeof(temp)) break;
+	  if ((size_t)(width + 2) > sizeof(temp)) break;
 
 	  sprintf(temp, tformat, va_arg(ap, void *));
